ARM: dts: Increase msm-core sampling poll
Polling every 50ms is too quick for the
MSM8992, leading to high loadavg.
However, polling values too high could
cause thermal adjustments to occur too
slowly, leading the device to heat up
unnecessarily. This is especially
dangerous for the MSM8992 as it has the
tendency to be damaged when heated and
enter the so-called “Boot Loop Of Death”.

Change-Id: I2e636871e0e11159be420c64b4d856372f89d49e
Signed-off-by: Diab Neiroukh <officiallazerl0rd@gmail.com>
Signed-off-by: mortezadadgar <mortezadadgar97@gmail.com>